OTA-server/requirements.txt
Flask>=1.1.0
cryptography>=3.3.1
python-dotenv>=0.15.0